% Generated by roxygen2: do not edit by hand
% Please edit documentation in litr_package_template.rmd.
\name{aw_covariates}
\alias{aw_covariates}
\title{Computing area weighted covariates (to be written)}
\usage{
aw_covariates(poly_in, poly_weight)
}
\arguments{
\item{poly_in}{}

\item{poly_weight}{}
}
\value{
A data.frame object
}
\description{
Computing area weighted covariates (to be written)
}
\examples{
# data

# run
aw_covariates()

}
\author{
Insang Song \email{geoissong@gmail.com}
}