% Generated by roxygen2: do not edit by hand
% Please edit documentation in litr_package_template.rmd.
\name{check_crs}
\alias{check_crs}
\title{Check Coordinate Reference System}
\usage{
check_crs(x)
}
\arguments{
\item{x}{sf/stars/SpatVector/SpatRaster object.}
}
\value{
A st_crs or crs object.
}
\description{
Check Coordinate Reference System
}
\examples{
# data
install.packages('spData')
library(spData)
data(us_states)
check_crs(us_states)

}
\author{
Insang Song \email{geoissong@gmail.com}
}