% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{check_crs2}
\alias{check_crs2}
\title{check_crs2: Coordinate system checker}
\usage{
check_crs2(input, crs_standard = "EPSG:4326")
}
\arguments{
\item{input}{Input object one of sf or terra::Spat* object}
}
\value{
A (reprojected) sf or SpatVector object.
}
\description{
The input is checked whether its coordinate system is present. If not, it is reprojected to EPSG:5179.
}