% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{clip_as_extent}
\alias{clip_as_extent}
\title{Extent clipping}
\usage{
clip_as_extent(pnts, buffer_r, nqsegs = NULL, target_input)
}
\arguments{
\item{pnts}{sf or SpatVector object}

\item{buffer_r}{numeric(1). buffer radius. this value will be automatically multiplied by 1.25}

\item{nqsegs}{integer(1). the number of points per a quarter circle; SOON TO BE DEPRECATED}

\item{target_input}{sf or SpatVector object to be clipped}
}
\value{
A clipped sf or SpatVector object.
}
\description{
Clip input vector by the expected maximum extent of computation.
}
\author{
Insang Song
}