% Generated by roxygen2: do not edit by hand
% Please edit documentation in litr_package_template.rmd.
\name{clip_as_extent_ras2}
\alias{clip_as_extent_ras2}
\title{clip_as_extent_ras2: Clip input raster (version 2).}
\usage{
clip_as_extent_ras2(points_in, buffer_r, nqsegs = 180, ras)
}
\arguments{
\item{points_in}{sf or SpatVector object}

\item{buffer_r}{numeric(1). buffer radius. this value will be automatically multiplied by 1.25}

\item{nqsegs}{integer(1). the number of points per a quarter circle}

\item{ras}{SpatRaster object to be clipped}
}
\description{
Clip input raster by the expected maximum extent of computation.
}
\author{
Insang Song
}