% Generated by roxygen2: do not edit by hand
% Please edit documentation in litr_package_template.rmd.
\name{clip_with}
\alias{clip_with}
\title{Clip with a polygon (to be written)}
\usage{
clip_with(x, poly, ...)
}
\arguments{
\item{x}{}

\item{poly}{}
}
\description{
Clip with a polygon (to be written)
}
\examples{
# data

# run
clip_with()

}
\author{
Insang Song \email{geoissong@gmail.com}
}