% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{distribute_process}
\alias{distribute_process}
\title{Process a given function in the entire or partial computational grids (under construction)}
\usage{
distribute_process(grids, grid_id = NULL, fun, ...)
}
\arguments{
\item{grids}{sf/SpatVector object. Computational grids.}

\item{grid_id}{character(1) or numeric(2). Default is NULL. If NULL, all grid_ids are used. "id_from:id_to" format or c(unique(grid_id)\link{id_from}, unique(grid_id)\link{id_to})}

\item{fun}{function supported in scomps.}

\item{...}{Arguments passed to fun.}
}
\value{
a data.frame object with mean value
}
\description{
Should
}
\author{
Insang Song \email{geoissong@gmail.com}
}