% Generated by roxygen2: do not edit by hand
% Please edit documentation in litr_package_template.rmd.
\name{extract_with_buffer.flat}
\alias{extract_with_buffer.flat}
\title{Subfunction: extract with buffers (flat weight; simple mean)}
\usage{
extract_with_buffer.flat(
  points,
  surf,
  radius,
  id,
  qsegs,
  kernel = NULL,
  bandwidth = NULL
)
}
\description{
Subfunction: extract with buffers (flat weight; simple mean)
}