% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{extract_with_buffer.kernel}
\alias{extract_with_buffer.kernel}
\title{Subfunction: extract with buffers (kernel weight; weighted mean)}
\usage{
extract_with_buffer.kernel(points, surf, radius, id, qsegs, kernel, bandwidth)
}
\description{
Subfunction: extract with buffers (kernel weight; weighted mean)
}