% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{extract_with_buffer}
\alias{extract_with_buffer}
\title{Extract summarized values from raster with points and a buffer radius (to be written)}
\usage{
extract_with_buffer(
  points,
  surf,
  radius,
  id,
  qsegs = 90,
  func = mean,
  kernel = NULL,
  bandwidth = NULL
)
}
\arguments{
\item{points}{SpatVector object. Coordinates where buffers will be generated}

\item{surf}{SpatRaster object. A raster of whatnot a summary will be calculated}

\item{radius}{numeric(1). Buffer radius. here we assume circular buffers only}

\item{id}{character(1). Unique identifier of each point.}

\item{qsegs}{integer(1). Number of vertices at a quarter of a circle. Default is 90.}

\item{func}{a function taking a numeric vector argument.}

\item{kernel}{character(1). Name of a kernel function (yet to be implemented)}

\item{bandwidth}{numeric(1). Kernel bandwidth.}
}
\value{
a data.frame object with mean value
}
\description{
For simplicity, it is assumed that the coordinate systems of the points and the raster are the same. Kernel function is not yet implemented.
}
\author{
Insang Song \email{geoissong@gmail.com}
}