% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{grid_merge}
\alias{grid_merge}
\title{grid_merge: Merge grid polygons with given rules}
\usage{
grid_merge(points_in, grid_in, grid_min_features)
}
\arguments{
\item{points_in}{sf or SpatVector object. Target points of computation.}

\item{grid_in}{sf or SpatVector object. The grid generated by sp_index_grid}

\item{grid_min_features}{integer(1). Threshold to merge adjacent grids.}
}
\value{
A sf or SpatVector object of computation grids.
}
\description{
Merge boundary-sharing (in "Rook" contiguity) grids with fewer target features than the threshold. This function strongly assumes that the input is returned from the sp_index_grid, which has 'CGRIDID' as the unique id field.
}
\examples{
# library(sf)
# library(igraph)
# ligrary(dplyr)
# dg = st_as_sfc(st_bbox(c(xmin = 0, ymin = 0, xmax = 8e5, ymax = 6e5)))
# st_crs(dg) = 5070
# dgs = st_as_sf(st_make_grid(dg, n = c(20, 15)))
# dgs$CGRIDID = seq(1, nrow(dgs))
#
# dg_sample = st_sample(dg, kappa = 5e-9, mu = 15, scale = 20000, type = "Thomas")
# st_crs(dg_sample) = st_crs(dg)
# dg_merged = grid_merge(st_as_sf(sss), dgs, 100)
#### NOT RUN ####
}
\author{
Insang Song
}