% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{initate_log}
\alias{initate_log}
\title{Turn on logging}
\usage{
initate_log(expr, dolog = FALSE, logpath)
}
\arguments{
\item{expr}{expression. Any function call to be logged.}

\item{dolog}{logical(1). Will the messages be logged.}

\item{logpath}{character(1). Log file path with the full log file name.}
}
\value{
Nothing. It will export a log file in the specified path as logpath.
}
\description{
Turn on logging
}
\author{
Insang Song
}