% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{switch_packbound}
\alias{switch_packbound}
\title{Switch spatial data class}
\usage{
switch_packbound(input)
}
\arguments{
\item{input}{Spat* in terra or sf object.}
}
\value{
Data converted to the other package class (if sf, terra; if terra, sf)
}
\description{
Convert stars into SpatRaster and vice versa; sf into SpatVector and vice versa.
}
\author{
Insang Song
}