% Generated by roxygen2: do not edit by hand
% Please edit documentation in litr_package_template.rmd.
\name{detect_class}
\alias{detect_class}
\title{Detect input classes}
\usage{
detect_class(x, template, ...)
}
\arguments{
\item{x}{A sf/stars/SpatVector/SpatRaster object.}

\item{template}{A sf/stars/SpatVector/SpatRaster object. Use this input to convert x into the template class.}
}
\value{
Class name or converted data object
}
\description{
Detect input classes
}
\examples{
# data

# run
detect_class()

}
\author{
Insang Song \email{geoissong@gmail.com}
}