% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{extract_with}
\alias{extract_with}
\title{A portal function to extract with points, buffers, and polygons (to be written)}
\usage{
extract_with(points, surf, radius, ...)
}
\arguments{
\item{points}{Points}

\item{surf}{Surface}

\item{radius}{Radius}
}
\value{
A data.frame object.
}
\description{
Extract with whatever sf or SpatVector objects are
}
\examples{
# data

# run
extract_with()

}
\author{
Insang Song \email{geoissong@gmail.com}
}